requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
pillow>=9.5.0
cloudinary==1.33.0
pydantic>=1.10.0,<2.0.0
python-multipart==0.0.6
uuid==1.30
python-dotenv==1.0.0
requests>=2.28.0
moviepy>=1.0.3
numpy>=1.21.0
